+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ## [Unreleased]
 
 ### Added
+- echance Markdown preview to support custom syntax
 ### Changed
 ### Deprecated
 ### Removed